requirements.txt
streamlit
scikit-learn
numpy
pandas
matplotlib.pyplot